Extract `IterationActor` from `BeamMobsim`
